requirements.txt
Flask==2.3.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
psycopg2
requests
celery
redis